feat: rename namespace and folder
